+++ src/commands/aircraft/takeoffPerf.ts
@@ -3,12 +3,14 @@ import { CommandCategory } from '../../constants';
 import { makeEmbed, makeLines } from '../../lib/embed';
 
 const takeoffPerfEmbed = makeEmbed({
-    title: 'FlyByWire A32NX | Where is the takeoff calculator?',
-    description: makeLines(['Currently, Airbus does not publicly release takeoff performance data for the A320neo, which means it is difficult to create an accurate takeoff calculator.',
+    title: 'FlyByWire A32NX | Takeoff Performance Calculator',
+    description: makeLines([
+        'A Takeoff Performance Calculator is available on the **Development** version of the A32NX. For more information, check out the [Takeoff Performance Calculator documentation](https://docs.flybywiresim.com/flypad-performance/).',
         '',
-        'While there are many online calculators available, these often use A320ceo data, and are not accurate.',
+        'For the **Stable** version, see the [V-Speeds Section](https://docs.flybywiresim.com/pilots-corner/beginner-guide/preparing-mcdu/#v-speeds) in our MCDU guide.',
         '',
-        'This unfortunately means that we cannot provide a reliable and accurate takeoff performance calculator at this time.']),
+        '**Note:** The functionality where the LSK next to the V1, VR and V2 speeds can be used to automatically calculate them in the MCDU PERF page has been removed in the Development version. Use the new Takeoff Performance Calculator instead.',
+    ]),
 });
 
 export const takeoffPerf: MessageCommandDefinition = {